Rename to containers, drop credstore_env, return container id
